Refactor expression function names for clarity
